cmd, k8s: new options for running pods with a specified scheduler name
Currently k8sscheduler tries to schedule every pod generated in a k8s
cluster. However, because of this behavior k8sscheduler conflicts with
the default scheduler. So this commit adds a new option -scheduler to
k8sscheduler: the value passed with the option is used as the field
scheduler.alpha.kubernetes.io/name of Pods' annotations. k8sscheduler
only schedules pods that have the name in the field.

This commit also adds a new option -scheduler to the podgen
command. If the command is specified, podgen creates pods with
specifying the scheduler with the annotations field.