cmd, k8s: more fine grained timers for getting unscheduled pods
Firmament is a scheduler that target sub-second placement latency, so
a unit of second for GetPodBatch() is too coarse grain. In addition,
an event of newly arrival pod can extend the timeout. For stopping
waiting new pods and making scheduling decision for low latency, this
commit adds a new option -ebpt to k8sscheduler. Even if new pods are
ready for scheduling, GetPodBatch() returns if the time exceeds the
timeout specified with the option.